requirements.txt
Django>=4.2
gunicorn>=20.1.0
psycopg2-binary
django-redis>=5.2.0
celery>=5.2.7 
python-decouple>=3.7 
django-import-export>=4.1.1
django-import-export[xlsx]
# whitenoise>=6.2.0 
pillow>=10.4.0
django-widget-tweaks>=1.0.5


